Moved types to separate files in order to minimize monster-large class
