+++ FastColoredTextBox/FastColoredTextBoxSupport.cs
@@ -1,0 +1,236 @@
+﻿//
+//  THIS CODE AND INFORMATION IS PROVIDED "AS IS" WITHOUT WARRANTY OF ANY
+//  KIND, EITHER EXPRESSED OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE
+//  IMPLIED WARRANTIES OF MERCHANTABILITY AND/OR FITNESS FOR A PARTICULAR
+//  PURPOSE.
+//
+//  License: GNU Lesser General Public License (LGPLv3)
+//
+//  Email: pavel_torgashov@ukr.net.
+//
+//  Copyright (C) Pavel Torgashov, 2011-2014. 
+using System;
+using System.Collections.Generic;
+using System.Drawing;
+using System.Text;
+
+namespace FastColoredTextBoxNS
+{
+   public enum WordWrapMode
+    {
+        /// <summary>
+        /// Word wrapping by control width
+        /// </summary>
+        WordWrapControlWidth,
+
+        /// <summary>
+        /// Word wrapping by preferred line width (PreferredLineWidth)
+        /// </summary>
+        WordWrapPreferredWidth,
+
+        /// <summary>
+        /// Char wrapping by control width
+        /// </summary>
+        CharWrapControlWidth,
+
+        /// <summary>
+        /// Char wrapping by preferred line width (PreferredLineWidth)
+        /// </summary>
+        CharWrapPreferredWidth,
+
+        /// <summary>
+        /// Custom wrap (by event WordWrapNeeded)
+        /// </summary>
+        Custom
+    }
+
+    public class PrintDialogSettings
+    {
+        public PrintDialogSettings()
+        {
+            ShowPrintPreviewDialog = true;
+            Title = "";
+            Footer = "";
+            Header = "";
+        }
+
+        public bool ShowPageSetupDialog { get; set; }
+        public bool ShowPrintDialog { get; set; }
+        public bool ShowPrintPreviewDialog { get; set; }
+
+        /// <summary>
+        /// Title of page. If you want to print Title on the page, insert code &amp;w in Footer or Header.
+        /// </summary>
+        public string Title { get; set; }
+
+        /// <summary>
+        /// Footer of page.
+        /// Here you can use special codes: &amp;w (Window title), &amp;D, &amp;d (Date), &amp;t(), &amp;4 (Time), &amp;p (Current page number), &amp;P (Total number of pages),  &amp;&amp; (A single ampersand), &amp;b (Right justify text, Center text. If &amp;b occurs once, then anything after the &amp;b is right justified. If &amp;b occurs twice, then anything between the two &amp;b is centered, and anything after the second &amp;b is right justified).
+        /// More detailed see <see cref="http://msdn.microsoft.com/en-us/library/aa969429(v=vs.85).aspx">here</see>
+        /// </summary>
+        public string Footer { get; set; }
+
+        /// <summary>
+        /// Header of page
+        /// Here you can use special codes: &amp;w (Window title), &amp;D, &amp;d (Date), &amp;t(), &amp;4 (Time), &amp;p (Current page number), &amp;P (Total number of pages),  &amp;&amp; (A single ampersand), &amp;b (Right justify text, Center text. If &amp;b occurs once, then anything after the &amp;b is right justified. If &amp;b occurs twice, then anything between the two &amp;b is centered, and anything after the second &amp;b is right justified).
+        /// More detailed see <see cref="http://msdn.microsoft.com/en-us/library/aa969429(v=vs.85).aspx">here</see>
+        /// </summary>
+        public string Header { get; set; }
+
+        /// <summary>
+        /// Prints line numbers
+        /// </summary>
+        public bool IncludeLineNumbers { get; set; }
+    }
+
+
+    /// <summary>
+    /// Type of highlighting
+    /// </summary>
+    public enum HighlightingRangeType
+    {
+        /// <summary>
+        /// Highlight only changed range of text. Highest performance.
+        /// </summary>
+        ChangedRange,
+
+        /// <summary>
+        /// Highlight visible range of text. Middle performance.
+        /// </summary>
+        VisibleRange,
+
+        /// <summary>
+        /// Highlight all (visible and invisible) text. Lowest performance.
+        /// </summary>
+        AllTextRange
+    }
+
+    /// <summary>
+    /// Strategy of search of end of folding block
+    /// </summary>
+    public enum FindEndOfFoldingBlockStrategy
+    {
+        Strategy1,
+        Strategy2
+    }
+
+    /// <summary>
+    /// Strategy of search of brackets to highlighting
+    /// </summary>
+    public enum BracketsHighlightStrategy
+    {
+        Strategy1,
+        Strategy2
+    }
+
+
+    public enum TextAreaBorderType
+    {
+        None,
+        Single,
+        Shadow
+    }
+
+    [Flags]
+    public enum ScrollDirection : ushort
+    {
+        None = 0,
+        Left = 1,
+        Right = 2,
+        Up = 4,
+        Down = 8
+    }
+
+    [Serializable]
+    public class ServiceColors
+    {
+        public Color CollapseMarkerForeColor { get; set; }
+        public Color CollapseMarkerBackColor { get; set; }
+        public Color CollapseMarkerBorderColor { get; set; }
+        public Color ExpandMarkerForeColor { get; set; }
+        public Color ExpandMarkerBackColor { get; set; }
+        public Color ExpandMarkerBorderColor { get; set; }
+
+        public ServiceColors()
+        {
+            CollapseMarkerForeColor = Color.Silver;
+            CollapseMarkerBackColor = Color.White;
+            CollapseMarkerBorderColor = Color.Silver;
+            ExpandMarkerForeColor = Color.Red;
+            ExpandMarkerBackColor = Color.White;
+            ExpandMarkerBorderColor = Color.Silver;
+        }
+    }
+
+#if Styles32
+    /// <summary>
+    /// Style index mask (32 styles)
+    /// </summary>
+    [Flags]
+    public enum StyleIndex : uint
+    {
+        None = 0,
+        Style0 = 0x1,
+        Style1 = 0x2,
+        Style2 = 0x4,
+        Style3 = 0x8,
+        Style4 = 0x10,
+        Style5 = 0x20,
+        Style6 = 0x40,
+        Style7 = 0x80,
+        Style8 = 0x100,
+        Style9 = 0x200,
+        Style10 = 0x400,
+        Style11 = 0x800,
+        Style12 = 0x1000,
+        Style13 = 0x2000,
+        Style14 = 0x4000,
+        Style15 = 0x8000,
+
+        Style16 = 0x10000,
+        Style17 = 0x20000,
+        Style18 = 0x40000,
+        Style19 = 0x80000,
+        Style20 = 0x100000,
+        Style21 = 0x200000,
+        Style22 = 0x400000,
+        Style23 = 0x800000,
+        Style24 = 0x1000000,
+        Style25 = 0x2000000,
+        Style26 = 0x4000000,
+        Style27 = 0x8000000,
+        Style28 = 0x10000000,
+        Style29 = 0x20000000,
+        Style30 = 0x40000000,
+        Style31 = 0x80000000,
+
+        All = 0xffffffff
+    }
+#else
+    /// <summary>
+    /// Style index mask (16 styles)
+    /// </summary>
+    [Flags]
+    public enum StyleIndex : ushort
+    {
+        None = 0,
+        Style0 = 0x1,
+        Style1 = 0x2,
+        Style2 = 0x4,
+        Style3 = 0x8,
+        Style4 = 0x10,
+        Style5 = 0x20,
+        Style6 = 0x40,
+        Style7 = 0x80,
+        Style8 = 0x100,
+        Style9 = 0x200,
+        Style10 = 0x400,
+        Style11 = 0x800,
+        Style12 = 0x1000,
+        Style13 = 0x2000,
+        Style14 = 0x4000,
+        Style15 = 0x8000,
+        All = 0xffff
+    }
+#endif
+}
